pyproject.toml
[project]
name = "quantixhackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "boto3>=1.40.68",
    "deepgram-sdk>=5.3.0",
    "faker>=37.12.0",
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic==2.7.4",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
]
